[Feat] 공격속도 추가
공격속도추가

+++ Assets/lcm/Scripts/Attack.cs
@@ -12,7 +12,7 @@ public class Attack : MonoBehaviour
     [SerializeField] float bulletSpeed;
 
     
-    public void Fire(int damage)
+    public void Fire(float damage)
     {
         GameObject instance = Instantiate(bulletPrefab, muzzlePoint.position, Quaternion.LookRotation(Vector3.forward));
         Rigidbody bulletRigidbody = instance.GetComponent<Rigidbody>();
